Moved all functions and data structures that are specific to the census data ic_type into a separate CensusData class, in a separate file. All of the functions previously in Initialization.H/cpp were specific to census data and have thus been moved into the CensusData class.
