+++ src/AgentContainer.cpp
@@ -6,29 +6,6 @@
 
 using namespace amrex;
 
-/*! \brief Assigns school by taking a random number between 0 and 100, and using
- *  default distribution to choose elementary/middle/high school. */
-AMREX_GPU_DEVICE AMREX_FORCE_INLINE
-int assign_school (const int nborhood, const amrex::RandomEngine& engine) {
-    int il4 = amrex::Random_int(100, engine);
-    int school = -1;
-
-    if (il4 < 36) {
-        school = 3 + (nborhood / 2);  /* elementary school */
-    }
-    else if (il4 < 68) {
-        school = 2;  /* middle school */
-    }
-
-    else if (il4 < 93) {
-        school = 1;  /* high school */
-    }
-    else {
-        school = 0;  /* not in school, presumably 18-year-olds or some home-schooled */
-    }
-    return school;
-}
-
 
 /*! Add runtime SoA attributes */
 void AgentContainer::add_attributes()
@@ -67,7 +44,7 @@ AgentContainer::AgentContainer (const amrex::Geometry            & a_geom,  /*!<
                                 0,
                                 RealIdx::nattribs,
                                 IntIdx::nattribs> (a_geom, a_dmap, a_ba),
-        student_counts(a_ba, a_dmap, SchoolType::total_school_type, 0)
+        m_student_counts(a_ba, a_dmap, SchoolType::total_school_type, 0)
 {
     BL_PROFILE("AgentContainer::AgentContainer");
 
@@ -75,7 +52,7 @@ AgentContainer::AgentContainer (const amrex::Geometry            & a_geom,  /*!<
     AMREX_ASSERT(m_num_diseases < ExaEpi::max_num_diseases);
     m_disease_names = a_disease_names;
 
-    student_counts.setVal(0);  // Initialize the MultiFab to zero
+    m_student_counts.setVal(0);  // Initialize the MultiFab to zero
 
     add_attributes();
 
@@ -126,428 +103,6 @@ AgentContainer::AgentContainer (const amrex::Geometry            & a_geom,  /*!<
 }
 
 
-/*! \brief Initialize agents for ExaEpi::ICType::Census
-
- *  + Define and allocate the following integer MultiFabs:
- *    + num_families: number of families; has 7 components, each component is the
- *      number of families of size (component+1)
- *    + fam_offsets: offset array for each family (i.e., each component of each grid cell), where the
- *      offset is the total number of people before this family while iterating over the grid.
- *    + fam_id: ID array for each family ()i.e., each component of each grid cell, where the ID is the
- *      total number of families before this family while iterating over the grid.
- *  + At each grid cell in each box/tile on each processor:
- *    + Set community number.
- *    + Find unit number for this community; specify that a part of this unit is on this processor;
- *      set unit number, FIPS code, and census tract number at this grid cell (community).
- *    + Set community size: 2000 people, unless this is the last community of a unit, in which case
- *      the remaining people if > 1000 (else 0).
- *    + Compute cumulative distribution (on a scale of 0-1000) of household size ranging from 1 to 7:
- *      initialize with default distributions, then compute from census data if available.
- *    + For each person in this community, generate a random integer between 0 and 1000; based on its
- *      value, assign this person to a household of a certain size (1-7) based on the cumulative
- *      distributions above.
- *  + Compute total number of agents (people), family offsets and IDs over the box/tile.
- *  + Allocate particle container AoS and SoA arrays for the computed number of agents.
- *  + At each grid cell in each box/tile on each processor, and for each component (where component
- *    corresponds to family size):
- *    + Compute percentage of school age kids (kids of age 5-17 as a fraction of total kids - under 5
- *      plus 5-17), if available in census data or set to default (76%).
- *    + For each agent at this grid cell and family size (component):
- *      + Find age group by generating a random integer (0-100) and using default age distributions.
- *        Look at code to see the algorithm for family size > 1.
- *      + Set agent position at the center of this grid cell.
- *      + Initialize status and day counters.
- *      + Set age group and family ID.
- *      + Set home location to current grid cell.
- *      + Initialize work location to current grid cell. Actual work location is set in
- *        ExaEpi::read_workerflow().
- *      + Set neighborhood and work neighborhood values. Actual work neighborhood is set
- *        in ExaEpi::read_workerflow().
- *      + Initialize workgroup to 0. It is set in ExaEpi::read_workerflow().
- *      + If age group is 5-17, assign a school based on neighborhood (#assign_school).
- *  + Copy everything to GPU device.
-*/
-void AgentContainer::initAgentsCensus (iMultiFab& num_residents,    /*!< Number of residents in each community (grid cell);
-                                                                         component 0: age under 5,
-                                                                         component 1: age group 5-17,
-                                                                         component 2: age group 18-29,
-                                                                         component 3: age group 30-64,
-                                                                         component 4: age group 65+,
-                                                                         component 4: total. */
-                                       iMultiFab& unit_mf,          /*!< Unit number of each community */
-                                       iMultiFab& FIPS_mf,          /*!< FIPS code (component 0) and
-                                                                         census tract number (component 1)
-                                                                         of each community */
-                                       iMultiFab& comm_mf,          /*!< Community number */
-                                       DemographicData& demo,       /*!< Demographic data */
-                                       const int nborhood_size      /*!< Size of neighborhood */ )
-{
-    BL_PROFILE("initAgentsCensus");
-
-    const Box& domain = Geom(0).Domain();
-
-    num_residents.setVal(0);
-    unit_mf.setVal(-1);
-    FIPS_mf.setVal(-1);
-    comm_mf.setVal(-1);
-
-    iMultiFab num_families(num_residents.boxArray(), num_residents.DistributionMap(), 7, 0);
-    iMultiFab fam_offsets (num_residents.boxArray(), num_residents.DistributionMap(), 7, 0);
-    iMultiFab fam_id (num_residents.boxArray(), num_residents.DistributionMap(), 7, 0);
-    num_families.setVal(0);
-
-    auto Nunit = demo.Nunit;
-    auto Ncommunity = demo.Ncommunity;
-    m_unit_teacher_counts_d.resize(Nunit, 0);
-    /* One can decide to define a iMultifab teachercounts -- but, data locality might be challenging*/
-    m_comm_teacher_counts_total_d.resize(Ncommunity, 0);
-    m_comm_teacher_counts_high_d.resize(Ncommunity, 0);
-    m_comm_teacher_counts_middle_d.resize(Ncommunity, 0);
-    m_comm_teacher_counts_elem3_d.resize(Ncommunity, 0);
-    m_comm_teacher_counts_elem4_d.resize(Ncommunity, 0);
-    m_comm_teacher_counts_daycr_d.resize(Ncommunity, 0);
-
-    amrex::Gpu::DeviceVector<long> student_teacher_ratios_d(m_student_teacher_ratios.size());
-    amrex::Gpu::copy(amrex::Gpu::hostToDevice, m_student_teacher_ratios.begin(), m_student_teacher_ratios.end(),
-                     student_teacher_ratios_d.begin());
-
-#ifdef AMREX_USE_OMP
-#pragma omp parallel if (Gpu::notInLaunchRegion())
-#endif
-    for (MFIter mfi(unit_mf, TilingIfNotGPU()); mfi.isValid(); ++mfi)
-    {
-        auto unit_arr = unit_mf[mfi].array();
-        auto FIPS_arr = FIPS_mf[mfi].array();
-        auto comm_arr = comm_mf[mfi].array();
-        auto nf_arr = num_families[mfi].array();
-        auto nr_arr = num_residents[mfi].array();
-
-        auto unit_on_proc = demo.Unit_on_proc_d.data();
-        auto Start = demo.Start_d.data();
-        auto FIPS = demo.FIPS_d.data();
-        auto Tract = demo.Tract_d.data();
-        auto Population = demo.Population_d.data();
-
-        auto H1 = demo.H1_d.data();
-        auto H2 = demo.H2_d.data();
-        auto H3 = demo.H3_d.data();
-        auto H4 = demo.H4_d.data();
-        auto H5 = demo.H5_d.data();
-        auto H6 = demo.H6_d.data();
-        auto H7 = demo.H7_d.data();
-
-        auto N5  = demo.N5_d.data();
-        auto N17 = demo.N17_d.data();
-        //auto N29 = demo.N29_d.data();
-        //auto N64 = demo.N64_d.data();
-        //auto N65plus = demo.N65plus_d.data();
-
-        auto ratios = student_teacher_ratios_d.dataPtr();
-        auto unit_teacher_counts_d_ptr = m_unit_teacher_counts_d.data();
-        auto comm_teacher_counts_total_d_ptr = m_comm_teacher_counts_total_d.data();
-        auto comm_teacher_counts_high_d_ptr = m_comm_teacher_counts_high_d.data();
-        auto comm_teacher_counts_middle_d_ptr = m_comm_teacher_counts_middle_d.data();
-        auto comm_teacher_counts_elem3_d_ptr = m_comm_teacher_counts_elem3_d.data();
-        auto comm_teacher_counts_elem4_d_ptr = m_comm_teacher_counts_elem4_d.data();
-        auto comm_teacher_counts_daycr_d_ptr = m_comm_teacher_counts_daycr_d.data();
-
-        auto bx = mfi.tilebox();
-        amrex::ParallelForRNG(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k, amrex::RandomEngine const& engine) noexcept
-        {
-            int community = (int) domain.index(IntVect(AMREX_D_DECL(i, j, k)));
-            if (community >= Ncommunity) { return; }
-            comm_arr(i, j, k) = community;
-
-            int unit = 0;
-            while (community >= Start[unit+1]) { unit++; }
-            unit_on_proc[unit] = 1;
-            unit_arr(i, j, k) = unit;
-            FIPS_arr(i, j, k, 0) = FIPS[unit];
-            FIPS_arr(i, j, k, 1) = Tract[unit];
-
-            int community_size;
-            if (Population[unit] < (1000 + DemographicData::COMMUNITY_SIZE * (community - Start[unit]))) {
-                community_size = 0;  /* Don't set up any residents; workgroup-only */
-            }
-            else {
-                community_size = DemographicData::COMMUNITY_SIZE;   /* Standard 2000-person community */
-            }
-
-            int p_hh[7] = {330, 670, 800, 900, 970, 990, 1000};
-            int num_hh = H1[unit] + H2[unit] + H3[unit] +
-                H4[unit] + H5[unit] + H6[unit] + H7[unit];
-            if (num_hh) {
-                p_hh[0] = 1000 * H1[unit] / num_hh;
-                p_hh[1] = 1000* (H1[unit] + H2[unit]) / num_hh;
-                p_hh[2] = 1000* (H1[unit] + H2[unit] + H3[unit]) / num_hh;
-                p_hh[3] = 1000* (H1[unit] + H2[unit] + H3[unit] + H4[unit]) / num_hh;
-                p_hh[4] = 1000* (H1[unit] + H2[unit] + H3[unit] +
-                                 H4[unit] + H5[unit]) / num_hh;
-                p_hh[5] = 1000* (H1[unit] + H2[unit] + H3[unit] +
-                                 H4[unit] + H5[unit] + H6[unit]) / num_hh;
-                p_hh[6] = 1000;
-            }
-
-            int npeople = 0;
-            while (npeople < community_size + 1) {
-                int il  = amrex::Random_int(1000, engine);
-
-                int family_size = 1;
-                while (il > p_hh[family_size]) { ++family_size; }
-                AMREX_ASSERT(family_size > 0);
-                AMREX_ASSERT(family_size <= 7);
-
-                nf_arr(i, j, k, family_size-1) += 1;
-                npeople += family_size;
-            }
-
-            AMREX_ASSERT(npeople == nf_arr(i, j, k, 0) +
-                         2*nf_arr(i, j, k, 1) +
-                         3*nf_arr(i, j, k, 2) +
-                         4*nf_arr(i, j, k, 3) +
-                         5*nf_arr(i, j, k, 4) +
-                         6*nf_arr(i, j, k, 5) +
-                         7*nf_arr(i, j, k, 6));
-
-            nr_arr(i, j, k, 5) = npeople;
-        });
-
-        int nagents;
-        int ncomp = num_families[mfi].nComp();
-        int ncell = num_families[mfi].numPts();
-        {
-            BL_PROFILE("setPopulationCounts_prefixsum")
-            const int* in = num_families[mfi].dataPtr();
-            int* out = fam_offsets[mfi].dataPtr();
-            nagents = Scan::PrefixSum<int>(ncomp*ncell,
-                            [=] AMREX_GPU_DEVICE (int i) -> int {
-                                int comp = i / ncell;
-                                return (comp+1)*in[i];
-                            },
-                            [=] AMREX_GPU_DEVICE (int i, int const& x) { out[i] = x; },
-                                               Scan::Type::exclusive, Scan::retSum);
-        }
-        {
-            BL_PROFILE("setFamily_id_prefixsum")
-            const int* in = num_families[mfi].dataPtr();
-            int* out = fam_id[mfi].dataPtr();
-            Scan::PrefixSum<int>(ncomp*ncell,
-                                 [=] AMREX_GPU_DEVICE (int i) -> int {
-                                     return in[i];
-                                 },
-                                 [=] AMREX_GPU_DEVICE (int i, int const& x) { out[i] = x; },
-                                 Scan::Type::exclusive, Scan::retSum);
-        }
-
-        auto offset_arr = fam_offsets[mfi].array();
-        auto fam_id_arr = fam_id[mfi].array();
-        auto& agents_tile = DefineAndReturnParticleTile(0, mfi);
-        agents_tile.resize(nagents);
-        auto aos = &agents_tile.GetArrayOfStructs()[0];
-        auto& soa = agents_tile.GetStructOfArrays();
-
-        auto age_group_ptr = soa.GetIntData(IntIdx::age_group).data();
-        auto family_ptr = soa.GetIntData(IntIdx::family).data();
-        auto home_i_ptr = soa.GetIntData(IntIdx::home_i).data();
-        auto home_j_ptr = soa.GetIntData(IntIdx::home_j).data();
-        auto work_i_ptr = soa.GetIntData(IntIdx::work_i).data();
-        auto work_j_ptr = soa.GetIntData(IntIdx::work_j).data();
-        auto hosp_i_ptr = soa.GetIntData(IntIdx::hosp_i).data();
-        auto hosp_j_ptr = soa.GetIntData(IntIdx::hosp_j).data();
-        auto nborhood_ptr = soa.GetIntData(IntIdx::nborhood).data();
-        auto school_ptr = soa.GetIntData(IntIdx::school).data();
-        auto workgroup_ptr = soa.GetIntData(IntIdx::workgroup).data();
-        auto work_nborhood_ptr = soa.GetIntData(IntIdx::work_nborhood).data();
-        auto random_travel_ptr = soa.GetIntData(IntIdx::random_travel).data();
-
-        int i_RT = IntIdx::nattribs;
-        int r_RT = RealIdx::nattribs;
-        int n_disease = m_num_diseases;
-
-        GpuArray<int*,ExaEpi::max_num_diseases> status_ptrs;
-        GpuArray<ParticleReal*,ExaEpi::max_num_diseases> counter_ptrs, timer_ptrs;
-        for (int d = 0; d < n_disease; d++) {
-            status_ptrs[d] = soa.GetIntData(i_RT+i0(d)+IntIdxDisease::status).data();
-            counter_ptrs[d] = soa.GetRealData(r_RT+r0(d)+RealIdxDisease::disease_counter).data();
-            timer_ptrs[d] = soa.GetRealData(r_RT+r0(d)+RealIdxDisease::treatment_timer).data();
-        }
-
-        auto dx = ParticleGeom(0).CellSizeArray();
-        auto my_proc = ParallelDescriptor::MyProc();
-
-        auto student_counts_arr = student_counts[mfi].array();
-
-        Long pid;
-#ifdef AMREX_USE_OMP
-#pragma omp critical (init_agents_nextid)
-#endif
-        {
-            pid = PType::NextID();
-            PType::NextID(pid+nagents);
-        }
-        AMREX_ALWAYS_ASSERT_WITH_MESSAGE(
-            static_cast<Long>(pid + nagents) < LastParticleID,
-            "Error: overflow on agent id numbers!");
-
-        amrex::ParallelForRNG(bx, ncomp, [=] AMREX_GPU_DEVICE (int i, int j, int k, int n, amrex::RandomEngine const& engine) noexcept
-        {
-            int nf = nf_arr(i, j, k, n);
-            if (nf == 0) return;
-
-            int unit = unit_arr(i, j, k);
-            int community = comm_arr(i, j, k);
-            int family_id_start = fam_id_arr(i, j, k, n);
-            int family_size = n + 1;
-            int num_to_add = family_size * nf;
-
-            int community_size;
-            if (Population[unit] < (1000 + DemographicData::COMMUNITY_SIZE * (community - Start[unit]))) {
-                community_size = 0;  /* Don't set up any residents; workgroup-only */
-            }
-            else {
-                community_size = DemographicData::COMMUNITY_SIZE;   /* Standard 2000-person community */
-            }
-
-            int p_schoolage = 0;
-            if (community_size) {  // Only bother for residential communities
-                if (N5[unit] + N17[unit]) {
-                    p_schoolage = 100*N17[unit] / (N5[unit] + N17[unit]);
-                }
-                else {
-                    p_schoolage = 76;
-                }
-            }
-
-            int start = offset_arr(i, j, k, n);
-            int nborhood = 0;
-            for (int ii = 0; ii < num_to_add; ++ii) {
-                int ip = start + ii;
-                auto& agent = aos[ip];
-                int il2 = amrex::Random_int(100, engine);
-                if (ii % family_size == 0) {
-                    nborhood = amrex::Random_int(DemographicData::COMMUNITY_SIZE / nborhood_size, engine);
-                }
-                int age_group = -1;
-
-                if (family_size == 1) {
-                    if (il2 < 28) { age_group = 4; }      /* single adult age 65+   */
-                    else if (il2 < 68) { age_group = 3; } /* age 30-64 (ASSUME 40%) */
-                    else { age_group = 2; }               /* single adult age 19-29 */
-                    nr_arr(i, j, k, age_group) += 1;
-                } else if (family_size == 2) {
-                    if (il2 == 0) {
-                        /* 1% probability of one parent + one child */
-                        int il3 = amrex::Random_int(100, engine);
-                        if (il3 < 2) { age_group = 4; }        /* one parent, age 65+ */
-                        else if (il3 < 62) { age_group = 3; }  /* one parent 30-64 (ASSUME 60%) */
-                        else { age_group = 2; }                /* one parent 19-29 */
-                        nr_arr(i, j, k, age_group) += 1;
-                        if (((int) amrex::Random_int(100, engine)) < p_schoolage) {
-                            age_group = 1; /* 22.0% of total population ages 5-18 */
-                        } else {
-                            age_group = 0;   /* 6.8% of total population ages 0-4 */
-                        }
-                        nr_arr(i, j, k, age_group) += 1;
-                    } else {
-                        /* 2 adults, 28% over 65 (ASSUME both same age group) */
-                        if (il2 < 28) { age_group = 4; }      /* single adult age 65+ */
-                        else if (il2 < 68) { age_group = 3; } /* age 30-64 (ASSUME 40%) */
-                        else { age_group = 2; }               /* single adult age 19-29 */
-                        nr_arr(i, j, k, age_group) += 2;
-                    }
-                }
-
-                if (family_size > 2) {
-                    /* ASSUME 2 adults, of the same age group */
-                    if (il2 < 2) { age_group = 4; }  /* parents are age 65+ */
-                    else if (il2 < 62) { age_group = 3; }  /* parents 30-64 (ASSUME 60%) */
-                    else { age_group = 2; }  /* parents 19-29 */
-                    nr_arr(i, j, k, age_group) += 2;
-
-                    /* Now pick the children's age groups */
-                    for (int nc = 2; nc < family_size; ++nc) {
-                        if (((int) amrex::Random_int(100, engine)) < p_schoolage) {
-                            age_group = 1; /* 22.0% of total population ages 5-18 */
-                        } else {
-                            age_group = 0;   /* 6.8% of total population ages 0-4 */
-                        }
-                        nr_arr(i, j, k, age_group) += 1;
-                    }
-                }
-
-                agent.pos(0) = (i + 0.5_rt)*dx[0];
-                agent.pos(1) = (j + 0.5_rt)*dx[1];
-                agent.id()  = pid+ip;
-                agent.cpu() = my_proc;
-
-                for (int d = 0; d < n_disease; d++) {
-                    status_ptrs[d][ip] = 0;
-                    counter_ptrs[d][ip] = 0.0_rt;
-                    timer_ptrs[d][ip] = 0.0_rt;
-                }
-                age_group_ptr[ip] = age_group;
-                family_ptr[ip] = family_id_start + (ii / family_size);
-                home_i_ptr[ip] = i;
-                home_j_ptr[ip] = j;
-                work_i_ptr[ip] = i;
-                work_j_ptr[ip] = j;
-                hosp_i_ptr[ip] = -1;
-                hosp_j_ptr[ip] = -1;
-                nborhood_ptr[ip] = nborhood;
-                work_nborhood_ptr[ip] = nborhood;
-                workgroup_ptr[ip] = 0;
-                random_travel_ptr[ip] = -1;
-
-                if (age_group == 0) {
-                    school_ptr[ip] = 5; // note - need to handle playgroups
-                } else if (age_group == 1) {
-                    school_ptr[ip] = assign_school(nborhood, engine);
-                } else {
-                    school_ptr[ip] = 0; // only use negative values to indicate school closed
-                }
-
-                // Increment the appropriate student counter based on the school assignment
-                if (school_ptr[ip] == SchoolType::elem_3) {
-                    amrex::Gpu::Atomic::AddNoRet(&student_counts_arr(i, j, k, SchoolType::elem_3), 1);
-                } else if (school_ptr[ip] == SchoolType::elem_4) {
-                    amrex::Gpu::Atomic::AddNoRet(&student_counts_arr(i, j, k, SchoolType::elem_4), 1);
-                } else if (school_ptr[ip] == SchoolType::middle) {
-                    amrex::Gpu::Atomic::AddNoRet(&student_counts_arr(i, j, k, SchoolType::middle), 1);
-                } else if (school_ptr[ip] == SchoolType::high) {
-                    amrex::Gpu::Atomic::AddNoRet(&student_counts_arr(i, j, k, SchoolType::high), 1);
-                } else if (school_ptr[ip] == SchoolType::day_care) {
-                    amrex::Gpu::Atomic::AddNoRet(&student_counts_arr(i, j, k, SchoolType::day_care), 1);
-                }
-
-                if (school_ptr[ip]>0) {amrex::Gpu::Atomic::AddNoRet(&student_counts_arr(i, j, k, SchoolType::total), 1); }
-
-            }
-        });
-
-        amrex::ParallelFor(bx,
-            [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
-        {
-            int comm = comm_arr(i,j,k);
-
-            comm_teacher_counts_high_d_ptr[comm]   = (int)((student_counts_arr(i, j, k, SchoolType::high))     / (ratios[SchoolType::high]));
-            comm_teacher_counts_middle_d_ptr[comm] = (int)((student_counts_arr(i, j, k, SchoolType::middle))   / (ratios[SchoolType::middle]));
-            comm_teacher_counts_elem3_d_ptr[comm]  = (int)((student_counts_arr(i, j, k, SchoolType::elem_3))   / (ratios[SchoolType::elem_3]));
-            comm_teacher_counts_elem4_d_ptr[comm]  = (int)((student_counts_arr(i, j, k, SchoolType::elem_4))   / (ratios[SchoolType::elem_4]));
-            comm_teacher_counts_daycr_d_ptr[comm]  = (int)((student_counts_arr(i, j, k, SchoolType::day_care)) / (ratios[SchoolType::day_care]));
-
-            int total = comm_teacher_counts_high_d_ptr[comm]
-                      + comm_teacher_counts_middle_d_ptr[comm]
-                      + comm_teacher_counts_elem3_d_ptr[comm]
-                      + comm_teacher_counts_elem4_d_ptr[comm]
-                      + comm_teacher_counts_daycr_d_ptr[comm];
-            comm_teacher_counts_total_d_ptr[comm] = total;
-            amrex::Gpu::Atomic::AddNoRet(&unit_teacher_counts_d_ptr[unit_arr(i,j,k,0)], total);
-        });
-    }
-
-    demo.CopyToHostAsync(demo.Unit_on_proc_d, demo.Unit_on_proc);
-    amrex::Gpu::streamSynchronize();
-}
 
 /*! \brief Return bin pointer at a given mfi, tile and model name */
 DenseBins<AgentContainer::PType>* AgentContainer::getBins (const std::pair<int,int>& a_idx,
@@ -704,7 +259,7 @@ void AgentContainer::moveAgentsToHome ()
 
     For each agent, set its position to a random location with a probabilty of 0.01%
 */
-void AgentContainer::moveRandomTravel (const iMultiFab& unit_mf)
+void AgentContainer::moveRandomTravel ()
 {
     BL_PROFILE("AgentContainer::moveRandomTravel");
