+++ src/IO.cpp
@@ -44,10 +44,7 @@ namespace IO
     + Write agents to file - see AgentContainer::WritePlotFile().
 */
 void writePlotFile (const AgentContainer& pc, /*!< Agent (particle) container */
-                    const iMultiFab& /*num_residents*/,
-                    const iMultiFab& unit_mf, /*!< MultiFab with unit number of each community */
-                    const iMultiFab& FIPS_mf, /*!< MultiFab with FIPS code and census tract ID */
-                    const iMultiFab& comm_mf, /*!< MultiFab of community number */
+                    const CensusData& censusData,  /*!< Contains census data */
                     const int num_diseases, /*!< Number of diseases */
                     const std::vector<std::string>& disease_names, /*!< Names of diseases */
                     const Real cur_time, /*!< current time */
@@ -62,9 +59,9 @@ void writePlotFile (const AgentContainer& pc, /*!< Agent (particle) container */
     output_mf.setVal(0.0);
     pc.generateCellData(output_mf);
 
-    amrex::Copy(output_mf, unit_mf, 0, ncomp_d*num_diseases  , 1, 0);
-    amrex::Copy(output_mf, FIPS_mf, 0, ncomp_d*num_diseases+1, 2, 0);
-    amrex::Copy(output_mf, comm_mf, 0, ncomp_d*num_diseases+3, 1, 0);
+    amrex::Copy(output_mf, censusData.unit_mf, 0, ncomp_d*num_diseases  , 1, 0);
+    amrex::Copy(output_mf, censusData.FIPS_mf, 0, ncomp_d*num_diseases+1, 2, 0);
+    amrex::Copy(output_mf, censusData.comm_mf, 0, ncomp_d*num_diseases+3, 1, 0);
 
     {
         Vector<std::string> plt_varnames = {};
@@ -180,10 +177,7 @@ void writePlotFile (const AgentContainer& pc, /*!< Agent (particle) container */
     + Sum across all processors and write to file.
 */
 void writeFIPSData (const AgentContainer& agents, /*!< Agents (particle) container */
-                    const iMultiFab& unit_mf, /*!< MultiFab with unit number of each community */
-                    const iMultiFab& /*FIPS_mf*/,
-                    const iMultiFab& /*comm_mf*/,
-                    const DemographicData& demo, /*!< Demographic data */
+                    const CensusData& censusData, /*!< Census data */
                     const std::string& prefix, /*!< Filename prefix */
                     const int num_diseases, /*!< Number of diseases */
                     const std::vector<std::string>& disease_names, /*!< Names of diseases */
@@ -209,7 +203,7 @@ void writeFIPSData (const AgentContainer& agents, /*!< Agents (particle) contain
         amrex::Print() << "Generating diagnostic data by FIPS code "
                        << "for " << disease_names[d] << "\n";
 
-        std::vector<amrex::Real> data(demo.Nunit, 0.0);
+        std::vector<amrex::Real> data(censusData.demo.Nunit, 0.0);
         amrex::Gpu::DeviceVector<amrex::Real> d_data(data.size(), 0.0);
         amrex::Real* const AMREX_RESTRICT data_ptr = d_data.dataPtr();
 
@@ -220,7 +214,7 @@ void writeFIPSData (const AgentContainer& agents, /*!< Agents (particle) contain
             {
                 for (MFIter mfi(*mf_vec[lev], TilingIfNotGPU()); mfi.isValid(); ++mfi)
                 {
-                    auto unit_arr = unit_mf[mfi].array();
+                    auto unit_arr = censusData.unit_mf[mfi].array();
                     auto cell_data_arr = (*mf_vec[lev])[mfi].array();
 
                     auto bx = mfi.tilebox();
