+++ src/main.cpp
@@ -10,7 +10,6 @@
 #include "AgentContainer.H"
 #include "CaseData.H"
 #include "DemographicData.H"
-#include "Initialization.H"
 #include "IO.H"
 #include "Utils.H"
 
@@ -33,7 +32,7 @@ void override_amrex_defaults ()
 int main (int argc, /*!< Number of command line arguments */
           char* argv[] /*!< Command line arguments */)
 {
-    amrex::Initialize(argc,argv,true,MPI_COMM_WORLD,override_amrex_defaults);
+    amrex::Initialize(argc, argv, true, MPI_COMM_WORLD, override_amrex_defaults);
 
     runAgent();
 
@@ -100,28 +99,21 @@ void runAgent ()
         amrex::Print() << "    " << params.disease_names[d] << "\n";
     }
 
-    DemographicData demo;
-    if (params.ic_type == ICType::Census) { demo.InitFromFile(params.census_filename, params.workgroup_size); }
-
     std::vector<CaseData> cases;
     cases.resize(params.num_diseases);
     for (int d = 0; d < params.num_diseases; d++) {
-        if (params.ic_type == ICType::Census && params.initial_case_type[d] == "file") {
+        if (params.initial_case_type[d] == "file") {
             cases[d].InitFromFile(params.disease_names[d], params.case_filename[d]);
         }
     }
 
-    Geometry geom = ExaEpi::Utils::get_geometry(demo, params);
-
+    Geometry geom;
     BoxArray ba;
     DistributionMapping dm;
-    ba.define(geom.Domain());
-    ba.maxSize(params.max_grid_size);
-    dm.define(ba);
-
-    amrex::Print() << "Base domain is: " << geom.Domain() << "\n";
-    amrex::Print() << "Max grid size is: " << params.max_grid_size << "\n";
-    amrex::Print() << "Number of boxes is: " << ba.size() << " over " << ParallelDescriptor::NProcs() << " ranks. \n";
+    CensusData censusData;
+    if (params.ic_type == ICType::Census) {
+        censusData.init(params, geom, ba, dm);
+    }
 
     // The default output filename is:
     // output.dat for a single disease
@@ -171,11 +163,6 @@ void runAgent ()
         }
     }
 
-    iMultiFab num_residents(ba, dm, 6, 0);
-    iMultiFab unit_mf(ba, dm, 1, 0);
-    iMultiFab FIPS_mf(ba, dm, 2, 0);
-    iMultiFab comm_mf(ba, dm, 1, 0);
-
     amrex::Vector< std::unique_ptr<MultiFab> > disease_stats;
     disease_stats.resize(params.num_diseases);
     for (int d = 0; d < params.num_diseases; d++) {
@@ -192,24 +179,12 @@ void runAgent ()
     {
         BL_PROFILE_REGION("Initialization");
         if (params.ic_type == ICType::Census) {
-            pc.initAgentsCensus(num_residents, unit_mf, FIPS_mf, comm_mf, demo, params.nborhood_size);
-            ExaEpi::Initialization::read_workerflow(demo, params, unit_mf, comm_mf, pc, params.workgroup_size);
+            censusData.initAgents(pc, params.nborhood_size);
+            censusData.read_workerflow(pc, params.workerflow_filename, params.workgroup_size);
             if (params.initial_case_type[0] == "file") {
-                ExaEpi::Initialization::setInitialCasesFromFile( pc,
-                                                                 unit_mf,
-                                                                 FIPS_mf,
-                                                                 comm_mf,
-                                                                 cases,
-                                                                 params.disease_names,
-                                                                 demo );
+                censusData.setInitialCasesFromFile(pc, cases, params.disease_names);
             } else {
-                ExaEpi::Initialization::setInitialCasesRandom(  pc,
-                                                                unit_mf,
-                                                                FIPS_mf,
-                                                                comm_mf,
-                                                                params.num_initial_cases,
-                                                                params.disease_names,
-                                                                demo );
+                censusData.setInitialCasesRandom(pc, params.num_initial_cases, params.disease_names);
             }
         } else if (params.ic_type == ICType::UrbanPop) {
             Abort("UrbanPop not yet implemented");
@@ -238,27 +213,11 @@ void runAgent ()
             amrex::Print() << "Simulating day " << i << " " << std::flush;
 
             if ((params.plot_int > 0) && (i % params.plot_int == 0)) {
-                ExaEpi::IO::writePlotFile(  pc,
-                                            num_residents,
-                                            unit_mf,
-                                            FIPS_mf,
-                                            comm_mf,
-                                            params.num_diseases,
-                                            params.disease_names,
-                                            cur_time,
-                                            i);
+                ExaEpi::IO::writePlotFile(pc, censusData, params.num_diseases, params.disease_names, cur_time, i);
             }
 
             if ((params.aggregated_diag_int > 0) && (i % params.aggregated_diag_int == 0)) {
-                ExaEpi::IO::writeFIPSData(  pc,
-                                            unit_mf,
-                                            FIPS_mf,
-                                            comm_mf,
-                                            demo,
-                                            params.aggregated_diag_prefix,
-                                            params.num_diseases,
-                                            params.disease_names,
-                                            i );
+                ExaEpi::IO::writeFIPSData(pc, censusData, params.aggregated_diag_prefix, params.num_diseases, params.disease_names, i);
             }
 
             // Update agents' disease status
@@ -371,7 +330,7 @@ void runAgent ()
             }
 
             if ((params.random_travel_int > 0) && (i % params.random_travel_int == 0)) {
-                pc.moveRandomTravel(unit_mf);
+                pc.moveRandomTravel();
                 using SrcData = AgentContainer::ParticleTileType::ConstParticleTileDataType;
                 on_travel_pc.copyParticles(pc,
                                            [=] AMREX_GPU_HOST_DEVICE (const SrcData& src, int ip) {
@@ -422,26 +381,11 @@ void runAgent ()
     }
 
     if (params.plot_int > 0) {
-        ExaEpi::IO::writePlotFile(  pc,
-                                    num_residents,
-                                    unit_mf,
-                                    FIPS_mf,
-                                    comm_mf,
-                                    params.num_diseases,
-                                    params.disease_names,
-                                    cur_time,
-                                    params.nsteps);
+        ExaEpi::IO::writePlotFile(pc, censusData, params.num_diseases, params.disease_names, cur_time, params.nsteps);
     }
 
     if ((params.aggregated_diag_int > 0) && (params.nsteps % params.aggregated_diag_int == 0)) {
-        ExaEpi::IO::writeFIPSData(  pc,
-                                    unit_mf,
-                                    FIPS_mf,
-                                    comm_mf,
-                                    demo,
-                                    params.aggregated_diag_prefix,
-                                    params.num_diseases,
-                                    params.disease_names,
-                                    params.nsteps);
+        ExaEpi::IO::writeFIPSData(pc, censusData, params.aggregated_diag_prefix, params.num_diseases,
+                                  params.disease_names, params.nsteps);
     }
 }
